@@ -32,41 +32,40 @@ dependencies:
     sdk: flutter
 
   # CSV
-  csv: ^5.0.2
+  csv: ^6.0.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  flutter_svg: ^0.22.0
-  intl: ^0.17.0
+  flutter_svg: ^2.2.3
+  intl: ^0.20.2
 
   # Google Sign In
-  firebase_core: ^2.24.2
-  firebase_auth: ^4.6.3
-  google_sign_in: ^6.1.4
+  firebase_core: ^4.4.0
+  firebase_auth: ^6.1.4
+  google_sign_in: ^7.2.0
 
   # Google Icon
-  font_awesome_flutter: ^9.0.0
+  font_awesome_flutter: ^10.12.0
 
   # State Management
   provider: ^6.1.1
   google_maps_flutter: ^2.5.0
-  permission_handler: ^11.0.1
-  location: ^5.0.3
-  firebase_database: ^10.3.8
+  permission_handler: ^12.0.1
+  location: ^8.0.1
+  firebase_database: ^12.1.2
   # firebase: ^9.0.3
-  firebase_storage: ^11.5.6
+  firebase_storage: ^13.0.6
   # flutter_local_notifications: ^8.2.0
-  timezone: ^0.8.0
   auto_size_text: ^3.0.0
   bottom_sheet: ^4.0.0
   calendar_view: ^1.0.4
   get: ^4.6.5
-  package_info_plus: ^4.2.0
+  package_info_plus: ^9.0.0
   sqflite: ^2.3.0
   flutter_smart_dialog: ^4.9.6
-  syncfusion_flutter_calendar: ^20.3.47
+  syncfusion_flutter_calendar: ^32.2.3
   shared_preferences: ^2.2.2
-  fluttertoast: ^8.2.4
+  fluttertoast: ^9.0.0
   table_calendar: ^3.0.8
   flutter_colorpicker: ^1.1.0
 
@@ -79,7 +78,7 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^6.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
